refactor: replaced direct database call with API-based script fetching
